+++ plugins/node-js/jparse-code-cli-tests.js
@@ -1,13 +1,13 @@
-import JParserToolsCli = require("./jparser-tools-cli");
-
+"use strict";
+var JParseCodeCli = require("./jparse-code-cli");
 /** Simple test for jparser-tools-cli
  * @author TeamworkGuy2
  * @since 2016-06-21
  */
-module JParserToolsCliTest {
-
-    export function testJsCli() {
-        var opts = JParserToolsCli.createOptions({
+var JParseCodeCliTest;
+(function (JParseCodeCliTest) {
+    function testJsCli() {
+        var opts = JParseCodeCli.createOptions({
             sources: [
                 { path: "src/services", depth: 1, fileExtensions: ["java"] },
                 { path: "src/models", depth: 2, fileExtensions: ["java"] },
@@ -17,14 +17,15 @@ module JParserToolsCliTest {
                 { path: "../models/model-defs.json", namespaces: ["App.Models", "App.Core"] },
             ],
             log: "../output/jparser-tools.log",
+            threads: 5,
+            debug: true
         });
-
-        var optsStr = JParserToolsCli.stringifyOptions(opts, '"');
-
+        var optsStr = JParseCodeCli.stringifyOptions(opts, '"');
         var expStr = ' -sources "' + "src/services=1,[java]" + ";" + "src/models=2,[java]" + '"' +
             ' -destinations "' + "../models/service-defs.json=[App.Services]" + ";" + "../models/model-defs.json=[App.Models,App.Core]" + '"' +
-            ' -log "' + "../output/jparser-tools.log" + '"';
-
+            ' -log "' + "../output/jparser-tools.log" + '"' +
+            ' -threads ' + 5 +
+            ' -debug';
         if (expStr !== optsStr) {
             throw new Error("options string mismatch, expected: \n'" + expStr + "'\nactual: \n'" + optsStr + "'");
         }
@@ -32,15 +33,10 @@ module JParserToolsCliTest {
             console.log("success 'testJsCli()'");
         }
     }
-
-}
-
-
+    JParseCodeCliTest.testJsCli = testJsCli;
+})(JParseCodeCliTest || (JParseCodeCliTest = {}));
 function main() {
-    JParserToolsCliTest.testJsCli();
+    JParseCodeCliTest.testJsCli();
 }
-
 main();
-
-
-export = JParserToolsCliTest;
+module.exports = JParseCodeCliTest;