+++ plugins/node-js/jparse-code-cli-tests.ts
@@ -1,13 +1,13 @@
-"use strict";
-var JParserToolsCli = require("./jparser-tools-cli");
+import JParseCodeCli = require("./jparse-code-cli");
+
 /** Simple test for jparser-tools-cli
  * @author TeamworkGuy2
  * @since 2016-06-21
  */
-var JParserToolsCliTest;
-(function (JParserToolsCliTest) {
-    function testJsCli() {
-        var opts = JParserToolsCli.createOptions({
+module JParseCodeCliTest {
+
+    export function testJsCli() {
+        var opts = JParseCodeCli.createOptions({
             sources: [
                 { path: "src/services", depth: 1, fileExtensions: ["java"] },
                 { path: "src/models", depth: 2, fileExtensions: ["java"] },
@@ -16,12 +16,19 @@ var JParserToolsCliTest;
                 { path: "../models/service-defs.json", namespaces: ["App.Services"] },
                 { path: "../models/model-defs.json", namespaces: ["App.Models", "App.Core"] },
             ],
-            log: "../output/jparser-tools.log"
+            log: "../output/jparser-tools.log",
+            threads: 5,
+            debug: true
         });
-        var optsStr = JParserToolsCli.stringifyOptions(opts, '"');
+
+        var optsStr = JParseCodeCli.stringifyOptions(opts, '"');
+
         var expStr = ' -sources "' + "src/services=1,[java]" + ";" + "src/models=2,[java]" + '"' +
             ' -destinations "' + "../models/service-defs.json=[App.Services]" + ";" + "../models/model-defs.json=[App.Models,App.Core]" + '"' +
-            ' -log "' + "../output/jparser-tools.log" + '"';
+            ' -log "' + "../output/jparser-tools.log" + '"' +
+            ' -threads ' + 5 +
+            ' -debug';
+
         if (expStr !== optsStr) {
             throw new Error("options string mismatch, expected: \n'" + expStr + "'\nactual: \n'" + optsStr + "'");
         }
@@ -29,10 +36,15 @@ var JParserToolsCliTest;
             console.log("success 'testJsCli()'");
         }
     }
-    JParserToolsCliTest.testJsCli = testJsCli;
-})(JParserToolsCliTest || (JParserToolsCliTest = {}));
-function main() {
-    JParserToolsCliTest.testJsCli();
+
 }
+
+
+function main() {
+    JParseCodeCliTest.testJsCli();
+}
+
 main();
-module.exports = JParserToolsCliTest;
+
+
+export = JParseCodeCliTest;